Check in latest changes

+++ caiman-test/programs/saxpy/array_get.wgsl
@@ -1,7 +1,7 @@
 [[group(0), binding(0)]]
 var<storage, read> a : array<f32, 256>;
 [[group(0), binding(1)]]
-var<storage, read> i : u32;
+var<storage, read> i : u64;
 [[group(0), binding(2)]]
 var<storage, write> v : f32;
